backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
sqlalchemy
alembic
openai
beautifulsoup4
lxml
python-multipart